Rename to splitkit
